requirements.txt
requests==2.28.1
google-cloud-bigquery==3.3.5
google-api-core==2.11.0
google-auth==2.11.0